+++ src/main/java/org/javlo/logger/JavloLogFactory.java
@@ -1,0 +1,11 @@
+package org.javlo.logger;
+
+import java.util.Map;
+
+import org.apache.commons.collections.map.HashedMap;
+
+public class JavloLogFactory {
+	
+	private final static Map<String, JavloLogger> logger = new HashedMap();
+
+}