+++ src/main/java/org/javlo/logger/JavloLogger.java
@@ -1,0 +1,11 @@
+package org.javlo.logger;
+
+public class JavloLogger {
+	
+	private String key;
+	
+	JavloLogger(String key) {
+		this.key = key;
+	}
+
+}