+++ src/main/webapp/modules/content/js/content_codemirror.js
@@ -1,0 +1,12 @@
+ jQuery(window).load(function() {
+	jQuery(".text-editor").each(function(index ) {
+		CodeMirror.fromTextArea(this, {
+			lineNumbers: true,
+			mode: jQuery(this).data("mode"),
+			foldGutter: true,
+			matchBrackets: true,
+			extraKeys: {"Ctrl-Space": "autocomplete"},
+			gutters: ["CodeMirror-linenumbers", "CodeMirror-foldgutter"]
+		});
+	});
+});